+++ cli/src/config.ts
@@ -1,10 +1,19 @@
-import { pathExists, readFile, readJSON, writeFile, writeJSON } from '@ionic/utils-fs';
 import Debug from 'debug';
+import { pathExists, readFile, readJSON, writeFile, writeJSON } from 'fs-extra';
 import { dirname, extname, join, relative, resolve } from 'path';
 
 import c from './colors';
 import { parseApkNameFromFlavor } from './common';
-import type { AndroidConfig, AppConfig, CLIConfig, Config, ExternalConfig, IOSConfig, WebConfig } from './definitions';
+import type {
+  AndroidConfig,
+  AppConfig,
+  CLIConfig,
+  Config,
+  ExternalConfig,
+  IOSConfig,
+  WebConfig,
+  XcodeExportMethod,
+} from './definitions';
 import { OS } from './definitions';
 import { fatal, isFatal } from './errors';
 import { logger } from './log';
@@ -267,7 +276,12 @@ async function loadIOSConfig(rootDir: string, extConfig: ExternalConfig): Promis
   const podPath = lazy(() => determineGemfileOrCocoapodPath(rootDir, platformDirAbs, nativeProjectDirAbs));
   const webDirAbs = lazy(() => determineIOSWebDirAbs(nativeProjectDirAbs, nativeTargetDirAbs, nativeXcodeProjDirAbs));
   const cordovaPluginsDir = 'capacitor-cordova-ios-plugins';
-
+  const buildOptions = {
+    exportMethod: extConfig.ios?.buildOptions?.exportMethod as XcodeExportMethod,
+    xcodeSigningStyle: extConfig.ios?.buildOptions?.signingStyle,
+    signingCertificate: extConfig.ios?.buildOptions?.signingCertificate,
+    provisioningProfile: extConfig.ios?.buildOptions?.provisioningProfile,
+  };
   return {
     name,
     minVersion: '14.0',
@@ -287,6 +301,7 @@ async function loadIOSConfig(rootDir: string, extConfig: ExternalConfig): Promis
     webDir: lazy(async () => relative(platformDirAbs, await webDirAbs)),
     webDirAbs,
     podPath,
+    buildOptions,
   };
 }